[design]홈 글자 위치 수정
