+++ src/styles/globals.css
@@ -367,6 +367,7 @@
       -webkit-font-smoothing: antialiased;
       -moz-osx-font-smoothing: grayscale;
       text-rendering: optimizeLegibility;
+      overflow-x: hidden;
    }
 
    button,
